(search) Fix script loading for mobile support
