+++ code/common/renderer/src/main/java/nu/marginalia/renderer/MustacheRenderer.java
@@ -10,10 +10,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import spark.Response;
 
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.OutputStreamWriter;
-import java.io.Writer;
+import java.io.*;
+import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Map;
 
@@ -51,19 +49,10 @@ public class MustacheRenderer<T> {
         return template.apply(model);
     }
 
-    private Writer getWriter(Response response) throws IOException {
-
-        // response.raw() has a getWriter() method that fits here, but this is a trap, as subsequent
-        // calls to response.raw().getOutputStream() will fail with an IllegalStateException; and we
-        // have internal code that does this.
-
-        return new OutputStreamWriter(response.raw().getOutputStream());
-    }
-
     @SneakyThrows
     public Object renderInto(Response response, T model) {
 
-        template.apply(model, getWriter(response));
+        response.raw().getOutputStream().write(template.apply(model).getBytes(StandardCharsets.UTF_8));
 
         return "";
     }
@@ -76,21 +65,14 @@ public class MustacheRenderer<T> {
     }
 
     @SneakyThrows
-    public <T2> void renderInto(Response response, T model, String name, List<T2> children) {
-        Context ctx = Context.newBuilder(model).combine(name, children).build();
-
-        template.apply(ctx, getWriter(response));
-    }
-
-    @SneakyThrows
-    public <T2> String render(T model, Map<String, ?> children) {
+    public String render(T model, Map<String, ?> children) {
         Context ctx = Context.newBuilder(model).combine(children).build();
         return template.apply(ctx);
     }
 
     @SneakyThrows
-    public <T2> void renderInto(Response response, T model, Map<String, ?> children) {
+    public void renderInto(Response response, T model, Map<String, ?> children) {
         Context ctx = Context.newBuilder(model).combine(children).build();
-        template.apply(ctx, getWriter(response));
+        response.raw().getOutputStream().write(template.apply(ctx).getBytes(StandardCharsets.UTF_8));
     }
 }
