+++ code/services-application/search-service/src/main/resources/templates/search/parts/search-form.hdb
@@ -2,9 +2,6 @@
     <div id="search-box">
         <h1>
             Search The Internet
-            <button id="mcfeast" aria-controls="#filters">
-                Filters
-            </button>
         </h1>
         <input type="text" name="query" placeholder="Search..." value="{{query}}" autocomplete="off">
